erp_app/requirements.txt
streamlit
sqlalchemy
pandas
plotly
passlib
bcrypt
python-dotenv
psycopg2-binary
openpyxl
faker
